Module claw::parse

Re-imported from claw_parse.

Modules

ast

Contains the abstract syntax tree definitions.

parser

Contains the logic of the actual parser, written using [chumsky].

Functions

build_report

Transforms a result to use [ariadne]’s [Report].

create_parser

Creates the main [Parser] to be used when parsing a Claw script.

get_source

Returns the [Source] of a string reference.

parse

Parses a str into AST.

Type Definitions

AST

Public type used to represent a complete program in AST form.

ParseError

Error type returned by the parsers.