Module claw::verify

Constants

SB3_DEFINITIONS

Shared definitions between multiple schemas, though right now just SB3_SCHEMA.

SB3_SCHEMA

The schema that a project.json file is checked against.

Functions

verify

The main function that verifies a [serde_json::Value] is a valid project JSON.

verify_string

Conveniance function that verifies the string form of a project JSON.